move constants.js to blog/lib dir
