+++ src/features/blog/hooks/usePost.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { API_BASE_URL } from "../../../config/constants";
+import { API_BASE_URL } from "../lib/constants";
 
 function usePost(id) {
   const [post, setPost] = useState(null);
