+++ src/features/blog/lib/constants.js
@@ -1,0 +1,2 @@
+export const API_BASE_URL =
+  "https://1hmfpsvto6.execute-api.ap-northeast-1.amazonaws.com/dev";
